requirements.txt
tensorflow>=2.10.0
numpy>=1.19.5
Pillow>=8.0.0
pytest>=7.0.0